relocate overview samples and update paths
